feat(infra): implement Owner/Patient JPA entities & repositories with Flyway and tests
- Add OwnerEntity, PatientEntity and JPA repositories
- Add Flyway V1 baseline with constraints and indexes
- Add DataJpa tests (H2 PG mode) and POJO tests
- Achieve >=80% coverage on infra entities

+++ backend/src/main/java/com/vetflow/api/infrastructure/persistence/entity/OwnerEntity.java
@@ -33,7 +33,7 @@ public class OwnerEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "name", length = 100, nullable = false)
     private String name;
@@ -66,8 +66,8 @@ public class OwnerEntity {
     }
 
     // Getters/setters (o usa Lombok si quieres en infra):
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = (name != null ? name.trim() : null); }
     public String getPhone() { return phone; }
